requirements.txt
fastapi==0.115.5
Pillow==10.4.0
pymodbus==3.8.6
smbus2==0.4.3
uvicorn==0.32.0
